+++ Sources/MessagingInApp/Gist/Managers/LogManager.swift
@@ -1,8 +1,10 @@
+import CioInternalCommon
 import Foundation
 
 class LogManager {
     let siteId: String
     let dataCenter: String
+    let gistQueueNetwork: GistQueueNetwork = DIGraphShared.shared.gistQueueNetwork
 
     init(siteId: String, dataCenter: String) {
         self.siteId = siteId
@@ -12,8 +14,12 @@ class LogManager {
     func logView(message: Message, userToken: String?, completionHandler: @escaping (Result<Void, Error>) -> Void) {
         do {
             if let queueId = message.queueId, let userToken = userToken {
-                try GistQueueNetwork(siteId: siteId, dataCenter: dataCenter, userToken: userToken)
-                    .request(LogEndpoint.logUserMessageView(queueId: queueId), completionHandler: { response in
+                try gistQueueNetwork.request(
+                    siteId: siteId,
+                    dataCenter: dataCenter,
+                    userToken: userToken,
+                    request: LogEndpoint.logUserMessageView(queueId: queueId),
+                    completionHandler: { response in
                         switch response {
                         case .success(let (_, response)):
                             if response.statusCode == 200 {
@@ -24,10 +30,16 @@ class LogManager {
                         case .failure(let error):
                             completionHandler(.failure(error))
                         }
-                    })
+                    }
+                )
             } else {
-                try GistQueueNetwork(siteId: siteId, dataCenter: dataCenter)
-                    .request(LogEndpoint.logMessageView(messageId: message.messageId), completionHandler: { response in
+                try gistQueueNetwork.request(
+                    siteId: siteId,
+                    dataCenter: dataCenter,
+                    userToken: nil,
+
+                    request: LogEndpoint.logMessageView(messageId: message.messageId),
+                    completionHandler: { response in
                         switch response {
                         case .success(let (_, response)):
                             if response.statusCode == 200 {
@@ -38,7 +50,8 @@ class LogManager {
                         case .failure(let error):
                             completionHandler(.failure(error))
                         }
-                    })
+                    }
+                )
             }
         } catch {
             completionHandler(.failure(error))