+++ Tests/MessagingInApp/MessagingInAppIntegrationTests.swift
@@ -51,7 +51,7 @@ class MessagingInAppIntegrationTest: IntegrationTest {
         navigateToScreen(screenName: "Home")
 
         let givenMessages = [
-            Message(messageId: "welcome-banner", campaignId: .random, pageRule: "Home")
+            Message(messageId: "welcome-banner", campaignId: .random, pageRule: "^(Home)$")
         ]
 
         let expectToBeginAnimation = expectation(description: "Begin animation")
@@ -75,12 +75,12 @@ class MessagingInAppIntegrationTest: IntegrationTest {
         navigateToScreen(screenName: "Home")
 
         let givenMessages = [
-            Message(messageId: "welcome-banner", campaignId: .random, pageRule: "Home")
+            Message(messageId: "welcome-banner", campaignId: .random, pageRule: "^(Home)$")
         ]
 
         onDoneFetching(messages: givenMessages)
         doneLoadingMessage(givenMessages[0])
-        XCTAssertEqual(currentlyShownModalMessage?.instanceId, givenMessages[0].instanceId)
+        XCTAssertEqual(currentlyShownModalMessage, givenMessages[0])
 
         navigateToScreen(screenName: "Settings")
 
@@ -223,6 +223,114 @@ class MessagingInAppIntegrationTest: IntegrationTest {
         XCTAssertNil(currentlyShownModalMessage)
         XCTAssertFalse(isCurrentlyLoadingMessage)
     }
+
+    // MARK: clearUserToken
+
+    // Code that runs when the profile is logged out of the SDK
+
+    func test_clearUserToken_givenModalMessageShown_givenModalHasPageRuleSet_expectDismissModal() {
+        navigateToScreen(screenName: "Home")
+
+        let givenMessages = [
+            Message(messageId: "welcome-banner", campaignId: .random, pageRule: "^(Home)$")
+        ]
+        onDoneFetching(messages: givenMessages)
+        doneLoadingMessage(givenMessages[0])
+        XCTAssertNotNil(currentlyShownModalMessage)
+
+        Gist.shared.clearUserToken()
+
+        XCTAssertNil(currentlyShownModalMessage)
+    }
+
+    func test_clearUserToken_givenModalMessageShown_givenModalHasNoPageRuleSet_expectDoNotDismissModal() {
+        navigateToScreen(screenName: "Home")
+
+        let givenMessages = [
+            Message(messageId: "welcome-banner", campaignId: .random, pageRule: nil)
+        ]
+        onDoneFetching(messages: givenMessages)
+        doneLoadingMessage(givenMessages[0])
+        XCTAssertNotNil(currentlyShownModalMessage)
+
+        Gist.shared.clearUserToken()
+
+        XCTAssertNotNil(currentlyShownModalMessage)
+    }
+
+    // The in-app SDK maintains a cache of messages that are returned from the backend. When a profile is logged out of the SDK, we expect the message cache is cleared otherwise we run the risk of displaying messages meant for profile A to profile B.
+    func test_clearUserToken_givenProfileLoggedOutAndNewProfileLoggedIn_expectLocalMessageCacheCleared() {
+        Gist.shared.setUserToken("profile-A")
+
+        XCTAssertTrue(Gist.shared.messageQueueManager.localMessageStore.isEmpty)
+        setupHttpResponse(code: 200, body: readSampleDataFile(subdirectory: "InAppUserQueue", fileName: "fetch_response.json").data)
+        Gist.shared.messageQueueManager.fetchUserMessages()
+        XCTAssertFalse(Gist.shared.messageQueueManager.localMessageStore.isEmpty)
+
+        // Expect no messages immediately after logging into another profile.
+        Gist.shared.clearUserToken()
+        XCTAssertTrue(Gist.shared.messageQueueManager.localMessageStore.isEmpty)
+        Gist.shared.setUserToken("profile-B")
+        XCTAssertTrue(Gist.shared.messageQueueManager.localMessageStore.isEmpty)
+
+        // Expect that after first fetch with new profile logged in, the message cache remains empty.
+        setupHttpResponse(code: 304, body: "".data)
+        Gist.shared.messageQueueManager.fetchUserMessages()
+        XCTAssertTrue(Gist.shared.messageQueueManager.localMessageStore.isEmpty)
+    }
+
+    // MARK: action buttons
+
+    func test_onCloseButton_expectShowNextMessageInQueue() throws {
+        navigateToScreen(screenName: "Home")
+
+        let givenMessages = [
+            Message(pageRule: "^(Home)$"),
+            Message(pageRule: "^(Home)$"),
+            Message(pageRule: nil)
+        ]
+
+        onDoneFetching(messages: givenMessages)
+        doneLoadingMessage(givenMessages[0])
+        XCTAssertEqual(currentlyShownModalMessage?.queueId, givenMessages[0].queueId)
+
+        onCloseActionButtonPressed()
+
+        doneLoadingMessage(givenMessages[1])
+
+        XCTAssertEqual(currentlyShownModalMessage?.queueId, givenMessages[1].queueId)
+
+        onCloseActionButtonPressed()
+
+        doneLoadingMessage(givenMessages[2])
+
+        XCTAssertEqual(currentlyShownModalMessage?.queueId, givenMessages[2].queueId)
+
+        onCloseActionButtonPressed()
+
+        XCTAssertNil(currentlyShownModalMessage)
+    }
+
+    func test_onCloseButton_givenNextMessageDoesNotMatchPageRule_expectDoNotShowNextMessageInQueue() throws {
+        navigateToScreen(screenName: "Home")
+
+        let givenMessages = [
+            Message(pageRule: "^(Home)$"),
+            Message(pageRule: "^(Settings)$") // expect to not show this message on close.
+        ]
+
+        onDoneFetching(messages: givenMessages)
+        doneLoadingMessage(givenMessages[0])
+        XCTAssertEqual(currentlyShownModalMessage?.queueId, givenMessages[0].queueId)
+
+        onCloseActionButtonPressed()
+
+        XCTAssertFalse(isCurrentlyLoadingMessage) // expect to not being loading a new message.
+
+        navigateToScreen(screenName: "Settings")
+
+        XCTAssertTrue(isCurrentlyLoadingMessage) // When page rule matches, we expect to load new message.
+    }
 }
 
 extension MessagingInAppIntegrationTest {
@@ -269,5 +377,11 @@ extension MessagingInAppIntegrationTest {
 
     func doneLoadingMessage(_ message: Message) {
         engineWebMock.underlyingDelegate?.routeLoaded(route: message.messageId)
+    }
+
+    func onCloseActionButtonPressed() {
+        // Triggering the close button from the web engine simulates the user tapping the close button on the in-app WebView.
+        // This behaves more like an integration test because we are also able to test the message manager, too.
+        engineWebMock.underlyingDelegate?.tap(name: "", action: GistMessageActions.close.rawValue, system: false)
     }
 }